% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigESS.R
\name{bigESS}
\alias{bigESS}
\title{Big ESS function}
\usage{
bigESS(sce, celltypes, aggmethod, specmethod)
}
\arguments{
\item{celltypes}{A vector of cell type labels}

\item{specmethod}{The method for specificity}

\item{x}{An sce object}

\item{addmethod}{The method for aggregating counts}
}
\value{
A either a matrix or a vector, depending on the method used
}
\description{
For a given sce, choose the cell types, aggregation method, and specificity method.  Return either vector or matrix
}
