% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau.R
\name{tau}
\alias{tau}
\title{Tau}
\usage{
tau(x)
}
\arguments{
\item{x}{A matrix of data}
}
\value{
A vector
}
\description{
Returns a vector of specificity scores using the Tau metric.
The main ESS function provides a wrapper for this to select
cell types and aggregation method
}
